Added Fixtures to conftest.py
