+++ tests/integration/test_kerchunk.py
@@ -1,6 +1,3 @@
-import logging
-import os
-import unittest
 from pathlib import Path
 
 import earthaccess
@@ -9,25 +6,6 @@ from fsspec.core import strip_protocol
 
 kerchunk = pytest.importorskip("kerchunk")
 pytest.importorskip("dask")
-
-logger = logging.getLogger(__name__)
-assertions = unittest.TestCase("__init__")
-
-assertions.assertTrue("EARTHDATA_USERNAME" in os.environ)
-assertions.assertTrue("EARTHDATA_PASSWORD" in os.environ)
-
-logger.info(f"Current username: {os.environ['EARTHDATA_USERNAME']}")
-logger.info(f"earthaccess version: {earthaccess.__version__}")
-
-
-@pytest.fixture(scope="module")
-def granules():
-    granules = earthaccess.search_data(
-        count=2,
-        short_name="SEA_SURFACE_HEIGHT_ALT_GRIDS_L4_2SATS_5DAY_6THDEG_V_JPL2205",
-        cloud_hosted=True,
-    )
-    return granules
 
 
 @pytest.mark.parametrize("protocol", ["", "file://"])
